Move Memoizer to internal util package
The Memoizer is only for internal use and not public.
